refactor: decouple DataCollectionConfigFactory and related classes from opennms-config
Move DataCollectionConfigFactory, DefaultDataCollectionConfigDao,
DataCollectionConfigParser, and DefaultResourceTypesDao to new
features/collection/config module. This frees the last daemon
(collectd) from opennms-config — all 13 daemon-boot modules are
now fully decoupled.

- Replace JaxbUtils.unmarshal() with Jackson XmlMapper in parser
- Remove dead init()/main() methods from DataCollectionConfigFactory
- Add backward-compat re-export in opennms-config POM
- Replace opennms-config with features/collection/config in
  daemon-boot-collectd POM (eliminates 30 lines of exclusions)
