+++ features/collection/config/src/main/java/org/opennms/netmgt/config/DataCollectionConfigFactory.java
@@ -1,0 +1,53 @@
+/*
+ * Licensed to The OpenNMS Group, Inc (TOG) under one or more
+ * contributor license agreements.  See the LICENSE.md file
+ * distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * TOG licenses this file to You under the GNU Affero General
+ * Public License Version 3 (the "License") or (at your option)
+ * any later version.  You may not use this file except in
+ * compliance with the License.  You may obtain a copy of the
+ * License at:
+ *
+ *      https://www.gnu.org/licenses/agpl-3.0.txt
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied.  See the License for the specific
+ * language governing permissions and limitations under the
+ * License.
+ */
+package org.opennms.netmgt.config;
+
+import org.opennms.netmgt.config.api.DataCollectionConfigDao;
+
+/**
+ * Static singleton holder for the {@link DataCollectionConfigDao}.
+ *
+ * <p>In Spring Boot daemons the instance is set via
+ * {@link #setInstance(DataCollectionConfigDao)} from the daemon configuration
+ * class. The legacy {@code init()} method that loaded via Spring BeanUtils
+ * has been removed — Karaf is dead.</p>
+ */
+public abstract class DataCollectionConfigFactory {
+
+    private static DataCollectionConfigDao m_singleton = null;
+
+    public static void setInstance(DataCollectionConfigDao instance) {
+        m_singleton = instance;
+    }
+
+    public static synchronized void reload() {
+        if (m_singleton == null)
+            throw new IllegalStateException("The factory has not been initialized");
+        m_singleton.reload();
+    }
+
+    public static synchronized DataCollectionConfigDao getInstance() {
+        if (m_singleton == null)
+            throw new IllegalStateException("The factory has not been initialized");
+        return m_singleton;
+    }
+}
