+++ features/collection/pom.xml
@@ -14,6 +14,7 @@
     <module>api</module>
     <module>client-rpc</module>
     <module>collectors</module>
+    <module>config</module>
     <module>core</module>
     <module>impl</module>
     <module>persistence-rrd</module>
